requirements.txt
fastapi==0.117.1
groq==0.31.1
joblib==1.5.2
keras==2.15.0
locust==2.40.5
locust-cloud==1.27.0
numpy==1.26.4
pandas==2.3.2
python-dotenv==1.1.1
regex==2025.9.18
scikit-learn==1.7.2
sentence-transformers==5.1.0
tensorflow==2.15.0
tf_keras==2.15.1
torch==2.8.0
transformers==4.56.2